refactor: Better files cohesion
