+++ src/Breakdown/MetricLabelsList/behaviours/fixLegendForUnspecifiedLabelValueBehavior.ts
